Create README.md
Looking for practical ways for investment professionals to use the data in the Investment Data Alliance (IDA) - https://www.cfainstitute.org/en/research/foundation/investment-data-alliance - to become better at ethically managing client's assets. Have SBBI and MSCI datasets available (to members) and list the free World Federation of Exchanges and Damodaran datasets. Will be listing Factset data, as well as Calcbench and Smartkarma services soon. Interested in anything related to any of this to build out the usefulness of the IDA.